Move custom-fields to components
